refactor(frontend): move common/ui/search components into shared/ui and features/home/ui (FR-006 US2.1)
