+++ frontend/src/app/(marketing)/about/components/AboutHero.tsx
@@ -1,4 +1,4 @@
-import { Icon } from '@/components/ui/Icon';
+import { Icon } from '@/shared/ui/Icon';
 import { APP_VERSION } from '@/lib/constants';
 
 /**
