+++ frontend/src/app/(marketing)/about/components/ContactSection.tsx
@@ -1,4 +1,4 @@
-import { Icon } from '@/components/ui/Icon';
+import { Icon } from '@/shared/ui/Icon';
 import { SOCIAL_LINKS } from '@/lib/constants';
 
 /**
