+++ frontend/src/app/(marketing)/about/components/HowItWorks.tsx
@@ -1,4 +1,4 @@
-import { Icon } from '@/components/ui/Icon';
+import { Icon } from '@/shared/ui/Icon';
 
 const features = [
   {
