+++ frontend/src/app/(marketing)/about/components/MissionCard.tsx
@@ -1,4 +1,4 @@
-import { Icon } from '@/components/ui/Icon';
+import { Icon } from '@/shared/ui/Icon';
 
 /**
  * "Our Mission" card.
