+++ frontend/src/components/podcasts/PodcastCard.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import Link from 'next/link';
-import { Card } from '@/components/ui/Card';
-import { Icon } from '@/components/ui/Icon';
+import { Card } from '@/shared/ui/Card';
+import { Icon } from '@/shared/ui/Icon';
 import type { Podcast } from '@/lib/api';
 
 interface PodcastCardProps {
