+++ frontend/src/shared/ui/Badge.tsx
@@ -1,5 +1,5 @@
 import { HTMLAttributes } from 'react';
-import { cn } from '@/lib/utils';
+import { cn } from './cn';
 
 export interface BadgeProps extends HTMLAttributes<HTMLDivElement> {
     variant?: 'primary' | 'secondary' | 'outline' | 'ghost';
