+++ frontend/src/shared/ui/index.ts
@@ -1,10 +1,14 @@
 /**
  * T068 — shared/ui barrel.
+ *
+ * US2 (005-frontend-coverage-cleanup) added `Badge`, `LoadingSpinner`, and
+ * `FullPageLoading` (merged from the deleted `components/ui/` legacy).
  */
 
 export { cn } from "./cn";
 export { Button, type ButtonProps, type ButtonVariant, type ButtonSize } from "./Button";
 export { Input, type InputProps } from "./Input";
 export { Card, CardHeader, CardBody, CardFooter } from "./Card";
-export { Spinner, Skeleton } from "./Spinner";
+export { Spinner, Skeleton, LoadingSpinner, FullPageLoading } from "./Spinner";
 export { Icon, type IconProps } from "./Icon";
+export { Badge, type BadgeProps } from "./Badge";
